fix: use centralized types from types.ts for setup wizard
Move HardwareInfo, SystemTools, SetupProfile type definitions to
types.ts and re-export from index.ts. Update hardware.ts, setup-wizard.ts,
and their tests to import from types.js instead of defining locally.
