+++ src/gemmaclaw/provision/index.ts
@@ -4,6 +4,13 @@ export { createLlamaCppManager } from "./llamacpp-manager.js";
 export { createGemmaCppManager } from "./gemmacpp-manager.js";
 export { downloadFile, execCommand, waitForHealthy, fileExists } from "./download.js";
 export { DEFAULT_MODELS, OLLAMA_RUNTIME, LLAMACPP_RUNTIME } from "./model-registry.js";
+export { detectHardware, detectSystemTools, formatHardwareInfo } from "./hardware.js";
+export {
+  createStdioWizardIO,
+  formatModelSize,
+  runAdvancedWizard,
+  selectQuickProfile,
+} from "./setup-wizard.js";
 export type {
   BackendId,
   ProvisionOpts,
@@ -11,6 +18,9 @@ export type {
   ProvisionResult,
   RuntimeHandle,
   RuntimeManager,
+  HardwareInfo,
+  SetupProfile,
+  SystemTools,
 } from "./types.js";
 export {
   ALL_BACKENDS,
